tests/requirements.txt
pytest==8.2.1
pytest-order==1.2.1
python-dotenv==1.0.1
requests==2.32.2
